fixed gy-blog page bug that caused image to not load properly. Restructured the project folders and got rid of unecessary files
