+++ components/common/content-container-with-hero-image.tsx
