+++ components/home/hero-image/hero-image.tsx
