+++ interfaces-and-types/_app/layout/interfaces-and-types.ts
@@ -1,0 +1,15 @@
+export interface INavContent {
+  dropdowns: IDropdown[]
+  navLinks: INavLink[]
+}
+
+export interface IDropdown {
+  dropdownTitle: string
+  mainDestination: string
+  dropdownOptions: INavLink[]
+}
+
+export interface INavLink {
+  linkText: string
+  destination: string
+}
