+++ lib/common/sanity/image-url-for.ts
@@ -1,0 +1,6 @@
+import sanity from './sanity'
+import imageUrlBuilder from '@sanity/image-url'
+
+const imageBuilder = imageUrlBuilder(sanity)
+
+export const imageUrlFor = (source: any) => imageBuilder.image(source)
