+++ next.config.js
@@ -2,6 +2,6 @@
 module.exports = {
   reactStrictMode: true,
   images: {
-    domains: ['cdn.sanity.io'],
+    domains: ['cdn.sanity.io', 'i.guim.co.uk'],
   },
 }
