+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import 'styles/globals.css'
 import type { AppProps } from 'next/app'
-import { Layout } from 'screens/_app/layout/layout'
+import { Layout } from 'components/_app/layout/layout'
 
 export function MyApp({ Component, pageProps }: AppProps) {
   return (
