+++ pages/consulting.tsx
@@ -1,3 +1,15 @@
-import { ConsultingPage } from 'screens/consulting/consulting'
+import { ContentContainerWithHeroImage } from 'components/common/content-container-with-hero-image'
+
+export function ConsultingPage() {
+  return (
+    <ContentContainerWithHeroImage
+      src="/images/consulting.jpeg"
+      mainHeading="Get Expert Preparation"
+      subHeading=""
+    >
+      <h1 className="text-7xl w-full text-center my-152">Some Page Content!</h1>
+    </ContentContainerWithHeroImage>
+  )
+}
 
 export default ConsultingPage
