+++ pages/homestay.tsx
@@ -1,3 +1,15 @@
-import { HomestayPage } from 'screens/homestay/homestay'
+import { ContentContainerWithHeroImage } from 'components/common/content-container-with-hero-image'
+
+export function HomestayPage() {
+  return (
+    <ContentContainerWithHeroImage
+      src="/images/homework-floor.jpeg"
+      mainHeading="A Support System"
+      subHeading=""
+    >
+      <h1 className="text-7xl w-full text-center my-152">Some Page Content!</h1>
+    </ContentContainerWithHeroImage>
+  )
+}
 
 export default HomestayPage
