--- screens/consulting/consulting.tsx
@@ -1,13 +1,0 @@
-import { ContentContainerWithHeroImage } from 'components/content-container-with-hero-image'
-
-export function ConsultingPage() {
-  return (
-    <ContentContainerWithHeroImage
-      src="/images/consulting.jpeg"
-      mainHeading="Get Expert Preparation"
-      subHeading=""
-    >
-      <h1 className="text-7xl w-full text-center my-152">Some Page Content!</h1>
-    </ContentContainerWithHeroImage>
-  )
-}
