--- screens/homestay/homestay.tsx
@@ -1,13 +1,0 @@
-import { ContentContainerWithHeroImage } from 'components/content-container-with-hero-image'
-
-export function HomestayPage() {
-  return (
-    <ContentContainerWithHeroImage
-      src="/images/homework-floor.jpeg"
-      mainHeading="A Support System"
-      subHeading=""
-    >
-      <h1 className="text-7xl w-full text-center my-152">Some Page Content!</h1>
-    </ContentContainerWithHeroImage>
-  )
-}
